requirements.txt
appdirs==1.4.3
astroid==2.3.2
backcall==0.1.0
boto==2.49.0
boto3==1.12.0
botocore==1.15.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
distlib==0.3.0
docutils==0.15.2
gensim==3.8.1
idna==2.8
ipykernel==5.1.3
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
jmespath==0.9.4
jupyter-client==5.3.4
jupyter-core==4.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.4.5
numpy==1.18.1
pandas==1.0.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
pylint==2.4.3
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
requests==2.22.0
s3transfer==0.3.3
scipy==1.4.1
six==1.12.0
smart-open==1.9.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.8
virtualenv==20.0.3
virtualenv-clone==0.5.3
wcwidth==0.1.8
wrapt==1.11.2
